+++ ValheimEnforcer/ValheimEnforcer.cs
@@ -20,7 +20,7 @@ namespace ValheimEnforcer
     {
         public const string PluginGUID = "MidnightsFX.ValheimEnforcer";
         public const string PluginName = "ValheimEnforcer";
-        public const string PluginVersion = "0.6.4";
+        public const string PluginVersion = "0.6.5";
 
         internal static ManualLogSource Log;
         internal ValConfig cfg;